# backend/src/auth/dtos/token.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty, IsString } from 'class-validator';

export class RequestTokenDto {
  @ApiProperty()
  @IsString()
  @IsNotEmpty()
  email: string;

  @ApiProperty()
  @IsString()
  @IsNotEmpty()
  password: string;
}

export class ResponseTokenDto {
  @ApiProperty()
  @IsString()
  @IsNotEmpty()
  access_token: string;
}